Create deploy web app
This script creates an EC2 instance in the us-west-2 region, installs Apache and serves a simple web page, and allows HTTP traffic to the instance using a security group. It also creates an Elastic IP address and associates it with the instance. The 'output' block prints the public IP address of the instance, which can be used to access the web page.